added api endpoint for get summary

+++ backend/server.js
@@ -80,6 +80,11 @@ app.post("/api/transactions", async (req, res) => {
 app.delete("/api/transactions/:id", async (req, res) => {
   try {
     const { id } = req.params;
+
+    if (isNaN(parseInt(id))) {
+      return res.status(404).json({ message: "Invalid id" });
+    }
+
     const result =
       await sql`DELETE FROM transactions WHERE id=${id} RETURNING *`;
 
@@ -92,6 +97,28 @@ app.delete("/api/transactions/:id", async (req, res) => {
       .json({ message: "Transaction deleted successfully" });
   } catch (error) {
     console.log("Error deleting the transaction", error);
+    res.status(500).json({ message: "Internal server error" });
+  }
+});
+
+//get total
+app.get("/api/transactions/summary/:userId", async (req, res) => {
+  try {
+    const { userId } = req.params;
+
+    const balanceResult =
+      await sql`SELECT COALESCE(SUM(amount),0) as balance FROM transactions WHERE user_id=${userId}`;
+    const incomeResult =
+      await sql`SELECT COALESCE(SUM(amount),0) as income FROM transactions WHERE user_id=${userId} AND amount>0`;
+    const expenseResult =
+      await sql`SELECT COALESCE(SUM(amount),0) as expenses FROM transactions WHERE user_id=${userId} AND amount<0`;
+    res.status(200).json({
+      balance: balanceResult[0].balance,
+      income: incomeResult[0].income,
+      expenses: expenseResult[0].expenses,
+    });
+  } catch (error) {
+    console.log("Error getting the summary", error);
     res.status(500).json({ message: "Internal server error" });
   }
 });
